site: improve demo load expierence
