+++ .dumi/theme/builtins/Previewer/CodePreviewer.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useRef, useState } from 'react';
 import { LinkOutlined, ThunderboltOutlined, UpOutlined } from '@ant-design/icons';
 import type { Project } from '@stackblitz/sdk';
 import stackblitzSdk from '@stackblitz/sdk';
-import { Alert, Badge, Flex, Tooltip } from 'antd';
+import { Badge, Flex, Tooltip } from 'antd';
 import { createStyles, css } from 'antd-style';
 import classNames from 'classnames';
 import { FormattedMessage, useLiveDemo, useSiteData } from 'dumi';
@@ -20,9 +20,7 @@ import ExternalLinkIcon from '../../icons/ExternalLinkIcon';
 import DemoContext from '../../slots/DemoContext';
 import SiteContext from '../../slots/SiteContext';
 import CodeBlockButton from './CodeBlockButton';
-import type { AntdPreviewerProps } from './Previewer';
-
-const { ErrorBoundary } = Alert;
+import type { AntdPreviewerProps } from '.';
 
 function compress(string: string): string {
   return LZString.compressToBase64(string)
@@ -321,13 +319,13 @@ createRoot(document.getElementById('container')).render(<Demo />);
   const stackblitzPrefillConfig: Project = {
     title: `${localizedTitle} - antd@${dependencies.antd}`,
     template: 'create-react-app',
-    dependencies:{
+    dependencies: {
       ...dependencies,
       react: '^19.0.0',
       'react-dom': '^19.0.0',
       '@types/react': '^19.0.0',
       '@types/react-dom': '^19.0.0',
-      '@ant-design/v5-patch-for-react-19': '^1.0.3'
+      '@ant-design/v5-patch-for-react-19': '^1.0.3',
     },
     description: '',
     files: {
@@ -353,11 +351,7 @@ createRoot(document.getElementById('container')).render(<Demo />);
   const codeBox: React.ReactNode = (
     <section className={codeBoxClass} id={asset.id}>
       <section className="code-box-demo" style={codeBoxDemoStyle} ref={demoContainer}>
-        {liveDemoNode || (
-          <ErrorBoundary>
-            <React.StrictMode>{previewDemo.current}</React.StrictMode>
-          </ErrorBoundary>
-        )}
+        {liveDemoNode || <React.StrictMode>{previewDemo.current}</React.StrictMode>}
       </section>
       {!simplify && (
         <section className="code-box-meta markdown">
